Introduction to bigdeal
Hello. I'm interested in building a set of tools 2 rapidly model and prototype
board games. I have tried making some of these tools before for individual games
that I have designed, only really tools to automatically generate assets given a
set of input data, so essentially an asset pipeline.

However one of the games that I'm interested in building (a game called Hydra)
has some very complex balance issues associated with probability and I've been
thinking since beginning the design for that game that what I really need to do
is simulate the game on a mass scale with computer players and look at the
statistical distribution of the results. For example I would like to be able to
simulate 1,000,000 games and show that within five sigma or something they all
finish within 20 minutes despite the fact that the outcome of the game is
entirely dependent on probability and hence the runtime is theoretically
infinite.

I could obviously just code up the tools and rules individually in a specific
programming language but I think there is a more general set of tools that we're
missing. This repository is my exploration into what those tools might look
like.

My reasoning about there being more general tools is existence of such things
for other domains such as video games and visual effects. Programmers are
notoriously bad at using visual tools, preferring text, but in more creative
industries they've shown more creativity in designing and building their own
tools.

My short list of tools that I think would be helpful for rapidly developing
board games is:

1. simulation tools to cheque balance runtime and other quantitative factors for
a wide set of parameters such as number of players and different starting setups
2. computer player testing tools allowing a simple interface for complex
computer players to be plugged in to existing games to test their efficacy
3. asset pipeline tools for the automatic creation of board game assets from
input data and output of those into, for example, printable forms or forms used
by tabletop simulator
4. Online web play tools to allow a group of players to play a game live whilst
also quickly being able to modify the rules without re compilation or
redeployment of web assets.

Those tools may seem quite distinct from each other but I believe that the
unifying element of them is having a computer understandable description of the
game with which to work. Essentially what they all need is a copy of the
rulebook that is a complete description of the game that is structured well
enough to be understandable and transformable by a machine.

As discussed above, one could simply try and represent board games in a
pre-existing language like Python. Clearly many people develop board games as
programmes and have success in doing so. However in order to meet all of those
above goals with the same artefact I believe we need a slightly different
representation. Python and associates are programming languages and in general
it's not possible to generate or understand a code without simply running it.

I think what is required is something more akin to a data description language
or markup –something that can be examined without necessarily requiring it's
logic to be executed. After all, when needing to build a deck of cards required
for many card games it is not necessary to understand the rules of any specific
game. The deck of cards is itself an object that stands alone without any
specific rules attached. My experience tells me but systems which consume data
rather than execute logic are generally systems the more general and simpler.

If I want these tools to be consumable by the sort of people that typically
design board games they also need to exist in the declarative role primarily in
a world that matches very closely to the way in which board game designers think
rather than the way in which the computer thinks. So at least I am targeting a
fourth order language that is declarative. My goal is the user who wants to use
the language to perhaps meet one of those above goals should also automatically
be able to get the benefits of all of the other tools for free without having to
restructure or redesign the game.

So I'm going to start by exploring this description language for board games and
I'm going to do that by trying to describe a wide variety of different games in
some kind of gradually developing pseudo language to understand what constructs
such a language would need to have. An important design goal is to make a
language that is completely tailored for describing board games with minimal
friction under for absolutely nothing else. The target is not a general purpose
tool. It is only for board games.

The first game is one called liars dice. I played this first with Zander at the
UK Board Game Expo '23 and then subsequently again with Zander and Elliott. It
is a dice based game with a relatively low number of moving parts. After all it
was played by pirates who are mostly drunk.

I have obviously been targeting a language that is minimal and consistent from
the start and hence trying to reuse constructs as much as possible. I started at
the top and generally worked my way down, so started by describing what a dice
is and then describing what the players have and then finally by describing the
rules as an imperative set of steps.

It feels like starting with a description of the objects in the game was a good
idea because it has landed me with a syntax of that is seemingly all about
sequences. A dice is sequence. The players are sequence. There is a construct in
the language which I'm calling right now a sequence reference the describes an
object in terms of taking it from another sequence. I think what has come out is
very declarative and very lazy. It seems to be well suited to describing
concepts in terms of sequences and filters on them and then subsequently
describing rules in something more akin to imperative language.

Part of the asset pipeline is also producing rules so one design goal I have is
being able to write human readable rules in line with the computer readable
rules which can then subsequently be converted into a document. As we know from
programming having comments in line is a powerful aid for understanding but
often the comments are treated as irrelevant to the computer. I want to take a
more Python like approach where the comments themselves are also objects in
language that can be inspected and transformed.

This is obviously not a consistent syntax yet and there is a lot more to work
out but I like the broad strokes of where this is ended up and I feel like a
language based on sequences works well with many common abstractions in board
games.