Ownership semantics; and example games Tic-tac-toe and Tomatomato
So far when using sequence references I have not had to be particularly clear
about how they actually work. The general principle I have been working with is
that writing `1 from sequence` acts as a description of how the variable gets
populated, and if this value is used will select you a random item from the
sequence.

This makes it easy to roll a die, it's just `1 from d6`. You get a random
element from the sequence `d6`, which remains unchanged. This is "sampling" the
sequence.

Here I have also been referring to `next 1 from deck`, the intention for which
is to sample cards not randomly but in order. So in that case there is some kind
of state about where in the sequence we are, and it's not clear yet where that
state lives. E.g. does it live on a per variable basis? Do I get a new cursor
every time I define a new variable? Or does it exist at the source, such that
everyone reading from a sequence is sharing a concept of `next`? And does it
make any difference to specify next in a type description, i.e. does it modify
some state which would be confusing later? Or is 'state' even necessary – do we
just need an understanding of order, and next always returns from the 'front'?
This is important because so far the declarations outside of imperative blocks
don't need to be executed in order, and that's a nice property to retain.

Even in Liar's Dice I made reference to the `next` item in a sequence, when
asking for the `next current_player`.

Another question is what "ordered" means. E.g. clearly `next 1 from deck` makes
sense but `next 1 from d6` is perhaps unintuitive – a dice is not ordered. This
is the first hint that there might be a distinction between types of sequence. I
think in this case we don't need any distinction on the implementation level –
if you want to draw the `next 12 from d6` and receive `1, 2, 3, 4, 5, 6, 1, 2,
3, 4, 5, 6` then so be it.

A deck is not an endless stream of cards – it is 52 cards, one of each. In the
cards example I have written:

	deck: every card

to represent this. This is distinct from `52 from card` because the 52 are
randomly sampled (so duplicates are expected), whereas `every` is pulling all
possible combinations, once. The intention is also different – we are saying the
deck is a finite collection rather than an endless sequence.

What has been unclear is how to handle finite collections that are expected to
be mutated. I think there is definitely a source of subtle ambiguities around
the lifetime of sequences. If a sequence only exists within a certain scope,
what happens to the objects in that sequence when it goes out of scope? In the
above example, if I don't move the cards in the hand when it goes out of scope,
what happens to them?

The wrong answer is "they just disappear". Things in board games don't ever just
disappear. The closest we come to is cards "being put back in the box" but this
is always made explicit in the rules. I think things that finitely exist in real
space should not be able to "get lost" and the rules should always be explicit
about what happens to them.

A slightly better answer is that they return to whence they came – the system
could keep track of where items originate from and return them there
automatically (I did include a `reset` keyword in an earlier draft of this
example). But that strikes me as a bit too magic and again prone to bugs. I
think the place I am settling on is explicit moves and ownership – sequences
that contain real space items must always be provably empty when they go out of
scope. I.e. this should be a problem:

	rules:>
		1 choices: take next 3 from deck
		add to hand of player: choose 1 to 3 from choices

Here I have written `take next X from deck` to mean "take the next 3 cards from
the deck and put them into this other sequence". Here we know there could be 1
to 2 cards left in the `choices` at the moment it goes out of scope. We should
have had to return it to the deck if it's not used.

But clearly, this is fine:

	rules:>
		rolls: 5 from d6
		add to score of player: sum of rolls

It's not a problem when `rolls` goes out of scope because these were just dice
rolls we sampled, and don't represent any real objects. So there is clearly a
distinction here between a sequence which contains only sampled items and a
sequence which contains "owned" items, or items that must be dealt with before
the sequence disappears. So far the only signal between the two is that one
sequence was populated with `take` whereas the other was populated without it.
Is this enough?

Consider:

	hand: 5 from deck

So far I have said that this describes both where the cards in the hand will
come from and that they are randomly sampled. But if I later write:

	add to hand: take 5 from deck

Is this a statement that makes sense? The hand was previously just a reference
to some cards from the deck, but now we are treating it as a finite collection
that we can add to. Again, does this matter?

If I write:

	5 card:
		value = 2
	twos: 5 from card where value = 2

Is there a distinction here between these two declarations, or are they
equivalent?

And finally, should I be able to `take 6 from d6` and have `d6` become an empty
set? Clearly as `d6` represents a dice this doesn't make a lot of sense… but
maybe that's not a concern of syntax.

Right now I appear to be following a rule that only sequences declared as a
specific numerical item can contain owned objects, e.g.

	1 deck: every card
	1 from deck
	'Allowed: sampling an owned sequence'
	take 1 from deck
	'Allowed: taking from an owned sequence'

	1 from card
	'Allowed: sampling an unowned sequence'
	take 1 from card
	'Rejected: taking from an unowned sequence'

So there is a distinction between "owning sequences" and "unowned sequences" and
mutations can only happen on "owning sequences". There is clearly more thinking
to do here around whether this distinction is necessary or whether any sequence
can be inferred to be owning or unowning depending on use of the `take` keyword.