Secret sequences; an example game of Hearts
This example initially ended up with a lot of nested loops as I iterated over
players making choices and taking turns. This was probably functional but there
was a lot of code to represent quite a simple concept, which was players taking
turns to make the same choice.

This was a code smell to me. The language so far is meant to be about sequences
and using them without iteration, so this probably means either the right
constructs aren't there or I'm not being encouraged to think in the right way.
This is especially concerning because there are large number of trick-taking
card games (e.g. whist) and so something so common as trick-taking should really
be easy to represent.

Eventually I landed on simplifying the loops to the following line:

	change played_card of other_players: take pick_response from other_players

This is saying, "each player picks a card". What's not clear here is the
ordering in which this happens, which in the explicit loop example was clear – a
player picks a card and then the next player picks a card.

At the same time, I was thinking of implementing 7 Wonders. There is a similar
action in 7 Wonders where each player picks a card and holds it face down until
all players have picked. The distinction here is that this action is performed
in parallel – each player can choose a card at the same time. The above line of
code would also be what I would write to represent this.

One way to deal with this distinction would be to just say "this line of code
always happens in parallel" and if you want specific ordering you need to do
looping. However, players picking one after the other is the more common case,
so we should optimise for that.

Another way would be to add some sort of modifier or keyword (like `in order` or
another `next`) to the sequence reference to disambiguate when the ordering
matters. However, I asked myself if it's possible to avoid more complexity in
the language through adding more words.

I asked myself why in 7 Wonders players can choose in parallel and in Hearts
players must choose in order when in BigDeal these operations appear to be the
same. The answer I realised is because of knowledge of game state: in 7 Wonders,
a card is selected from a private independent hand and kept private, whilst in
Hearts, a card is selected from a private independent hand and made public. This
distinction means that players that come later in the turn order have an
advantage because they can choose a card based on what other players have
already picked.

So the differentiator between these two situations is in if more public state is
revealed between each choice. If a choice is being made from a secret sequence
and then being placed into a secret sequence, player choices can execute in
parallel. If the choice is being made from a public or shared sequence (like a
deck) or from a private sequence into a public sequence (like a player card)
player choices must execute serially.

For this reason I've introduced the `secret` keyword. This specifies that
certain collections are hidden from players. The two examples are each player's
`hand: 0 secret card` which is only visible to the player that owns it because
it is defined within the player sequence, and then the global `1 secret deck`
which isn't visible to any players at all because it is defined outside.

The `secret` keyword is more generally required anyway. If playing a game in a
web browser with multiple players, clearly the game executor needs to know what
data players should be able to see and what data they should not see. If there
was no `secret` keyword, the executor wouldn't be able to distinguish public
sequences from private ones.

The fortuitous discovery is that the `secret` keyword also allows things to
happen in parallel, automatically, where possible, without the need for any more
information. The executor already can infer whether all players can choose at
once or if each player must take turns.