Open classes; an example game of cribbage
One of the features that I've been using consistently so far is the idea that
BigDeal has "open classes", i.e. that you can extend the fields on a object
after it has already been defined and even potentially created. This is proven
to be quite a useful feature already in the existing examples cause it allows a
really flexible roles system to emerge where a player can be temporarily given
extra properties or methods when they become e.g. the current player or the
dealer. Players then lose those properties or methods when they stop acting as
that role – the open classes are not just a permanent change but also a scoped
change. However this causes a slight problem.

So far when I've written a new role an added properties to it the semantics have
been that the properties defined on the set get translated onto all the objects
in set. In fact this is how it works even for defining a type in the first
place. A card is still a sequence and when you define fields on it you were
really defining the fields on the items in the sequence.

Indeed when you use a property of a sequence in many examples already you are in
fact just generating a new sequence of all the properties of all the items.
However there are already cases where we want to define a sequence based on the
overall sequence rather than on the items in the sequence. For example, `sum` –
in this example we are computing an aggregate property of the whole sequence and
not just accessing a property of the items in the sequence.

There are already more examples: `diagonals of grid` is only a thing that can be
computed in the context of an entire grid and not just the individual items.
Obviously all of these could just be built-in but it would be more consistent to
be able to define a property that only applies to the sequence itself and not to
the items.

I can almost think of a world where this doesn't need to be true and where e.g.
computing the diagonals of a grid might be possible just by referring to
neighbours and next items... but i can't quite think of how that would work in
some key cases.

There's more to this in fact because not only do we need to define sequences at
the sequence level we also need to define properties at the sequence level. For
example in the tic tac toe example we defined the number of rows and columns in
the grid and it was not expected that these would be something that is available
on the squares of the grid but is in fact a property of the grid itself.

I think the correct factoring is still one that by default transmits properties
onto the items in the sequence because that seems to be the common case in all
the examples that I've written so far. However there is clearly a need to be
able to define properties of sequences rather than items in order to define
objects that are extensible like a grid.

One possibility is that we could reuse the `of` syntax when defining new
properties to specify where the property is defined. By default it will be
defined on the source objects but we could also specify it is defined at a
specific level.

	1 board: grid
		many from blank
		rows of grid: 3
		cols of grid: 3

I'm not super clear on what the significance of the name following the `of` is
in this example: what would the difference be between `rows of grid` and `rows
of board`? That should presumably have some difference in the final structure.
If it does not then perhaps we don't need to have an identifier there at all and
can just say `rows of: 3`.

This example is also showing something new which is perhaps a mixed definition –
we are defining the contents of the grid and also properties on it which is not
something we have done thus far.

An alternative might be to define the methods on the sequence from outside:

	diagonals of grid: groups of grid where any:
		group_by = #forward_diagonal_index
		group_by = #backward_diagonal_index

	slice of grid:
		count: many
		next count from grid

	slices of grid: many slice of grid

This works quite well. Now we need to think about how to extend this to subsets
of collections, e.g.

	card where suit = ❤️:
	score = 1