Non-integer numbers
While coding games so far I've only found one use case for real numbers:
dividing a sequence into fractions based on the number of players. Based on
other games I've played I can only remember one other instance where non-whole
numbers are used, and that's for scoring "half points" or similar. So where they
are used, they are typically simple fractions close to zero on the number line.

I think this is easily explained through humans not generally having an
intuitive grasp of non-whole numbers beyond halves or thirds, and putting this
sort of thing into game rules is exclusionary to anyone who doesn't get on well
with maths.

So then we consider what support for non-integer numbers should exist in
BigDeal. The default for most programming languages is just to use IEEE floating
points as a built-in, I think for two reasons: 1) there is some support for them
at the processor level, so adding support in the language is required to take
advantage of that hardware performance, and 2) it is a general purpose
abstraction for representing either very large (1E99) or very small (1E-99)
numbers which permits a wide array of use cases. I could obviously use floating
point to represent halves, thirds etc.

However, there are a lot of issues with floating points, most notably that their
limited precision is not something that most humans are used to dealing with.
They behave unintuitively when operating on them – there is a famous Python or
Javascript example that 0.2 + 0.3 != 0.5. There is also a huge amount of
academic literature focused on how to understand propagation of and correct
errors in floating point math. Clearly if I'm targeting a non-technical user,
this sort of surprising result is not something I want to occur.

Also, the high range of values supported by floating point is not something we
need. Board games don't contain numbers in the millions, billions, micros or
nanos.

So instead I'm going to only support fractions natively, both as literals (1/2,
1/3) and as the result of division computations (1/count of players => 1/5).
There is already an example where saying `take 1/5 from deck` means `take count
of deck/5 from deck`, but there is an advantage of the system being able to work
directly with the fractional form because it can decide what to do when the
number of things in `deck` is not divisible directly.

The next question is what happens to division by zero. If you do 1/0 in floating
point math, you typically either get a program error or a +/- infinity quantity.
I cannot think of any instances where board game rules would expect a human to
actually divide by zero or would expect them to reason that such a division
would produce infinity. I think if such a thing ends up in a BigDeal
description, it is most likely an edge case that the rules haven't considered,
or a bug in the rules.

So the plan then is for BigDeal to actively reject instances where division by
zero is possible as early as possible – so essentially, make possible division
by zero a compile time error and avoid any possibility it can happen at runtime.

Certainly, writing a literal `1/0` should be rejected as invalid. More tricky is
where we write `1/variable` – BigDeal will need to work out the range of
`variable` and emit an error if it could possibly contain zero. It will also
need to handle the case where `variable` actually can be zero, but that this
condition has been explicitly checked elsewhere, e.g. this should not produce an
error:

    deck: 0 to many card
    when count of deck > 0:
        change something: take 1/count of deck from deck