Take semantics; an example game of Go Fish
I've been a bit worried since writing an (as yet incomplete) Yahtzee example
that the principles I've been following around ownership of objects are not well
founded. I had convinced myself that  objects using the `take` keyword needed to
be declared specially to indicate that they were "owned" objects (eg 1 X), and
this was a distinct kind of thing from "unowned" objects (eg 1 from X). I now
can't remember the rationale for introducing this (such is the price of not
writing devlogs consistently).

I wrote some justification in a previous log around objects that are "taken"
needing to be kept track of to ensure they don't accidentally go out of scope. I
still think this is significant and useful but by itself doesn't justify special
syntax for ownership.

So I have sat down to write the simplest game I can think of that involves a lot
of cards moving around: Go Fish. It turns out not to be such a trivial game!
There are a few intricacies that is has teased out, and not just with take
semantics.

Firstly, we'll start with the easy one. I had assumed that `take` (being a verb)
would be confined to imperative contexts only. However this is an example where
the initial conditions need to be described using `take`: we are dealing some of
the cards to players – quite common! But also I don't think this is problematic,
as with non-mutating references it's clear what the initial conditions for the
source sequence are, and so are available for parsing without mutation in assets
generation. So in general `take` can be allowed anywhere.

Secondly, if you take a generative sequence, does that take objects from the
source sequences? In Go Fish I wrote:

	add to books of players: take every book of players

where `book` is a generative sequence:

	book:
	        value: 1 from value of hand
	        cards: 4 from hand where value = value of book

The intention was to only create books that are possible, and then take the
associated cards out of the hand. But how does this work with the values?
Obviously they are a property of the cards in the hand so what does it mean to
take them?

This looks like a consistency issue because `value of card` is an equivalent
structure to `cards of book`. So expecting it to take `cards of book` and not
`value of hand` sounds hard. True, one is an explicit property and the other is
an implicit flattened sequence, but relying on this logic sounds like it is
exposing too many internals to the user.

Really this distinction is between value being "a property of a thing" and cards
being "a real thing", and `take` could only apply to "real things". My
discomfort mentioned earlier was about whether this distinction is important –
maybe this is a reason that it is? However this is also quite a niche case, and
so the impact of a user incorrectly specifying objects as real or not might not
be immediately apparent to them.

Alternatively, the rules can be as they are and I need clearer semantics about
how to use `take` in this case:

	book:
		value: 1 from value of hand
		cards: take 4 from hand where value = value of book

What does it mean to use `take` here? It is so far not something that I've used,
and I've sort of left it unspecified when a `take` actually gets applied,
because we've only used it in the imperative case so far where the semantics are
clearer. Also, so far sequence references have not needed to have a clear moment
of when their values are computed because this is purely an immutable operation
(although I suspect there is possible issue around needing to eagerly compute a
reference to preserve some state, but this hasn't come up so far).

But it feels possible to define `take` to mean "remove the items from the source
sequence when computing the value" and this should suffice, he says with hope,
and this would certainly work fine for this case. The executor computing the
sequence reference will need to understand that the source sequence is not
immutable and ensure that references using `take` will not use the same items
twice.

If we go with this for now, let's return to:

	add to books of players: every book of players

I've removed the `take` here now, but do we need it? (I think we're about to
start talking about pointers, which is really not something I want users to have
to think about.) When this is executed, it would take the cards from the hand
and they'd "live" in the book. The cards need to stay in scope which I think
implies the book needs to stay in scope unless, say, the cards were moved out of
it, and then it could die. So it sounds like a sequence which contains "owned"
objects also becomes implicitly "owned" itself (or I might start saying
"ownership tracked") until the owned objects it contains are removed.

So in that case the book "lives" in the "books of players". But is there a clear
need for a `take` here? Part of me says no because the items in the generative
sequence are just being generated on demand. But that is singling out generative
sequences for special behaviour, which would be bad. But there feels like an
issue here, because if I say:

	my_book: 1st from every book of players

then where are the cards living? Well, I suppose given the rule above, only the
"1st from every book" actually exists, and the cards for that book are the only
ones "taken", and I suppose they must now live in `my_book`. So here the
ownership has implicitly transferred… but is this desirable? Because writing:

	my_card: 1st from deck

clearly does not implicitly transfer any ownership. So now there is required
knowledge involved of whether or not the source sequence contains owned objects
– this feels unintuitive and hard to keep track of for the user. So it feels
like `take` must always be used when there is a transfer of ownership – that is
what it's there for after all!

There are perhaps two options for dealing with `1st from every book`; one is to
disallow sequences generated from `take` statements to be referenced without
`take` themselves i.e. `1st from every book` would not be allowed by itself. But
again this is singling out generative sequences because `1st from deck` is
clearly fine. So this doesn't sound great.

Two is not to process a `take` unless the parent object has an explicit `take`.
So saying `1st from every book` would not transfer ownership of cards from the
hand to the book, it would just "pre-generate" the object and provide a
reference to it, which could then later be ``take` if desired. This seems to
work? So we would have

	book:
		value: 1 from value of hand
		cards: take 4 from hand where value = value of book

	add to books of players: take every book of players

and this would transfer the ownership of the cards into the book, and ownership
of the book into the books.

And thus we have come full circle, because above I was asking "how does `take`
of a parent object know which things to take?" and we have arrived at an answer
to that question: it only takes things that are themselves marked with `take`!

This actually makes a lot of sense, because as we know a card is:

	card:
		suit: 1 from club, diamond, heart, spade
		value: 1 from 1 to 10, jack, queen king

but clearly `take 1 from deck` does not transfer ownership of the `suit` or the
`value`. So those sequences not defined with `take` are unaffected, and `take`
only applies to the aggregate `card`. But defining `take` on the cards in the
book means that they are affected when their parent is taken.

So thus, by something like a proof by contradiction, owned objects can include
ownership of other objects, and that ownership is transferred when the object is
transferred too. `take` is used in both cases to indicate what is owned, by
defining something as being removed from a source sequence.