Choose semantics; a badly-written example of Canasta
I've been working on an example game of Canasta for the past few weeks. This
example has proven to be challenging, but as with any difficult endeavour it has
also been fruitful for pointing out issues and exploring new ground.

Canasta is a game of many rules and edge cases – despite having played many tens
of games (against a computer) there were still some rules I didn't fully
understand. Ultimately the example is still a WIP and there are many parts of it
that are not satisfactory. I have accepted that it's too difficult for now
without more tools so it's time to move on and learn more from elsewhere first –
needing more experience to complete a difficult task is normal for a novice.

The semantics of the `choose` word is one feature that Canasta has forced
exploration of. This word defines where the player needs to make a choice, and
it is attached to a sequence reference (like `take`) to define what the
available options for the choice are. So `chosen: choose 1 from hand` defines
that chosen will be 1 card picked by the player from their hand.

The point of making it a word attached to a sequence is to allow clients to
present a suitable interface for making the choice. The principle is that the
player should only be able to make valid choices, ie there is only pre-filtering
of the options rather than post-validation of the choice.

So far the choices have been quite simple. In the example I define a meld
sequence to only contain valid melds based on the rules of what cards it can
contain. When writing the `choose` to pick cards to make a meld, I ran into a
number of difficulties.

The first was how to reuse in the `choose` the logic for what is a valid meld. I
considered a number of options. One that seemed promising was adding a new type
of `where` statement that would specify the intended destination of the choice
and reuse the logic from that sequence, e.g.

	cards: choose from hand where for cards of meld

This works ok for creating new objects but not so well for any other operation.
Choices could also be about picking items to remove from a sequence or inserting
in a specific place, so just saying "for" is not general enough.

I instead considered promoting `choose` to be a more general type of expression
and not just a sequence modifier, e.g.:

	choose add to cards of meld: many from hand

This makes it explicit that the choice is intended for a specific future
operation. But again this seems somewhat limited as it looks like it can only
handle choices that are bound to a single expression.

I realised two helpful things. First, the type of choice i'm trying to represent
here is actually already possible with the current choose expressions. Asking
whether something is a valid item is just equivalent to asking whether it exists
in the sequence of items:

	chosen_cards: choose many from hand where:
		1 from meld where cards:
			cards of existing_meld
			chosen_cards

This is pleasing but took me a while to realise so there's a risk this isn't
intuitive at all.

Secondly, I have been trying to get the user to be explicit but this isn't
necessarily required, and the system can infer validation rules in many cases.
E.g in this case:

	Cards: choose 1 to many from hand
	Meld: choose 1 from existing_melds
	Add to cards of meld: take cards

it's clear that the choice of the cards and the meld need together to result in
a valid meld, and there's no reason why the system can't infer this.

What's needed to make this work is clarity about how the `where` clause applies
to sequences. So far it has mostly been used to specify initial conditions, and
the usage of it outside of that has been somewhat unclear. In the previous
examples I have specified things like `hand: 6 from deck` mainly as
documentation. Similarly in order to specify initial conditions I have in some
places elided where clauses altogether – how to specify initial conditions needs
to be explored.

This is perhaps suggesting that `where` conditions need to apply for the whole
lifetime of the sequence, and adding objects to the sequence can only be done if
the items added meet the condition. These conditions can be back propagated to
filter the possible player choices. It would be a compile time error to specify
incompatible conditions. This looks promising but it might cause the rigidity of
type systems to read their ugliness. More exploration is needed.

Another hard `choose` case was in specifying initial melds. In Canasta players
have to play melds of at least a certain points value before they can do
anything else. Following the principle of pre-filtering, the system needs to
know if a large enough initial meld is possible before letting the players pick
cards, and then only allow them to pick cards that form into such a meld.

I thought for a long time about how to express this with various choose
expressions without post-validation, but the situation is complex. At the very
least it requires loops to allow the player to pick multiple cards and form
multiple melds, and somehow the system needs to be able to limit what those
melds are.

I did consider specifying the necessary validation in a loop with an exit
condition that ensures the point total is met, and then requiring
implementations to allow players to undo their choices so they can manually
backtrack. Even so, there's a lot of faffing about moving things and it all
detracts from the actual game rules. I think trying to represent this situation
would be challenging for the average user.

I realised that I was trying to use `choose` both with assumptions about how the
user interface would work and also to specify the intended way in how players
make choices (i.e. they pick cards first and the make melds from them). Instead,
I should be expressing the choice as closely to the rules as possible and
allowing implementations to decide how to present the choices. So for initial
melds I settled on:

	Initial_melds: choose 1 from combinations of melds where points of combinations >= requirement

This succinctly expresses the rule and allows the implementation some freedom in
how to present it. If it computes that only 1 or 2 combinations are viable, it
could present these, or it can always fall back to allowing the user to pick
cards first. In another example:

	Choose 1 meld:
		Value: 1 from value of card
		Cards: take 3 to 7 from hand where value = value of meld

Two ways to implement this from the rules are 1) the user will pick the value
first and then the cards, 2) the system will realise both values are "floating"
and make the first card define the value for the others. The correct place for
that is indeed in the implementation.

So the principles i'm settling on for now are:

	1. Sequence validation logic will be automatically back-tracked to choices –
	there is no need to be explicit unless the rules need to be more restrictive
	2. High-level choices are favoured over multiple low level ones
	3. How both of these things are actually done is defined by the implementation

Another area which requires more work is handling of "secret" because canasta
has a rule that only the top card of the discard pile is visible, which is not
something that it is currently clear how to implement. Thankfully I'm going to
look at Hanabi next where a core element of the gameplay is revealing
information.