Boards; a Flash Point example and Cluedo WIP
I originally started implementing Cluedo to explore secrecy semantics but it
turned out none of what was there advanced the state of the art. What it did
help me explore was games that have boards because all the examples up until now
have only been card-based.

Whilst implementing Cluedo I used a grid mechanism but didn't really have any
comprehension about how to model walls – in the model of the grid I had at the
beginning, every square was just connected to every other and so walls were not
really supported.

There were various models that were possible. Whilst thinking about other games
I also realised that lots of boards have some kind of "zones" which encompass
more than a single grid square. For example, some games have rooms in a building
and define rules based on if players are in the same room or not.

I figured that Cluedo was a board on a grid and I could describe the rooms as an
overlay on that grid, and then a client could pick out that overlay and display
it differently. I did describe the whole board for Cluedo using data structures
but it was quite an undertaking to get the shapes of the rooms correct, and even
then some rooms have shapes which don't fit onto a grid.

I did experiment with using box drawing characters to visually represent a board
as text, with the idea being that players could sketch their board using box
drawing characters and then the parser would convert it into an underlying
representation. This proved to be too difficult to wield because the intuition
is to represent one character as one square but we really need to represent
properties of edges as well, and it's impossible to line up the board of the
edges also have character width.

I realised that I've been focusing exclusively on users being able to edit a
textual representation of the board but this is not really one of my
requirements. A board is a visual construct and so the best tool for creating
one will also be visual.

I've thought about whether a textual representation of the board is valuable,
even if users aren't expected to edit it directly. I did consider whether we
could just have a binary representation of the board that users have to edit
with a specific tool.

I think the answer is that a text-based representation is still valuable even if
the users don't commonly edit it directly. At the moment we haven't needed any
splitting of files – even complex games have only ended up at a couple of
hundred lines. I'm loath to introduce this because it introduces a new class of
issues. For example, uploading a game to a web client would need to support a
whole package format rather than simple text. Plus, import systems typically
introduce attack surface during its interaction with a file system. So the plan
for now is to have a tool that can provide a visual editor for a board that
outputs text.

One of the goals is to have an asset pipeline that can start with a data
representation and generate graphics for a board ideally as automatically as
possible. To explore this further I decided to implement another game that
inspired this project, which is Flash Point. It has a complex board with
multiple layers of overlay and dynamic walls that alter paths through the board
as the game progresses.

For a start, I found use of 2D ranges to make things much simpler. This just
involves extending the `to` operator to compute a sequence based on all the
dynamic properties, making it really just a subset of `every` with a starting
and finishing item. You can now do fun things with this, such as computing a
range of cards with:

	3♣️ to 5❤️

Here I have also fleshed out some of the interface for the `grid`. The grid is
"just" a sequence of squares which is generated automatically based on the row
and col ranges. Each square has an `edges` property that is expected to return
an edge sequence. The default implementation returns a default edge for each
adjacent square, not including diagonals.  The `edge` interface just requires a
`src` and `dst` property that specifies the edge in terms of squares.

Calling `paths` on the grid returns a sequence of all possible paths along the
`edges` without cycles which can be filtered to just valid paths for the game.
(I have had this idea that `grid` is a standard library structure that can be
implemented purely in BigDeal, but I don't have any idea about how `paths` would
be implemented yet without some sort of recursive definition that hasn't been
explored. SQL has recursive querying that is worth exploring.)

This has resulted in a board representation that is not too terrible, although I
would still expect users to want a graphical editor.

## Turning towards graphics

An important consideration is that BigDeal is intended to have multiple clients,
some of which are not graphical. Even amongst the graphical ones there is
considerable variation about what output format they will require. So in general
the graphical portion of the game represents a binding of the generic rules to a
specific client. I can therefore free myself from having to think hard about
specific formats at this stage.

However, there does need to be some way for a board editor to save its work and
ideally in a format that clients just working with raw BigDeal can synthesise
their own board. The goal is for eg a terminal-based client to be able to work
with a textual set of rules and still present a reasonable UI automatically. So
some presentation information does need to be derivable from the text format,
and an editor will need to write some graphical information to one file and some
key things to the text.

Each of the overlays in Flash Point has ended up as a property of the square on
the grid. This might be a reasonable heuristic – each of the zones become
properties of the grid squares which become "layers" on the graphical board. An
asset generation system could compute the polygons of each layer and allow them
to be styled, as in the Flash Point board where the quadrants are styled with
dashed borders.

Another common construct is "spaces" i.e. a physical item in the game being
moved to a set space which is initially empty. An example might be the outlines
of a pile of cards that you see in card games, indicating where cards could be
dragged and dropped. The construct i've been using for a space is `0 to 1 from …
initially empty`, and I've not yet felt compelled to introduce any more sugar to
this. It seems reasonable that this could be identified as a pattern by clients
or emitted by editors and styled accordingly.

Despite trying to not think about formats, I think I am imagining SVGs as a good
choice for a board editor output. They are just XML which means we can store
whatever metadata we want alongside them including the game textual rules, and
they can also hold raster images to use as graphics for various layers. They are
also easily rendered into other popular formats. I am imagining using an SVG as
a package format to allow a full game with graphics to be easily portable.

There are also a fair few language tweaks in here, such as an implicit
assumption that `flashovers` is a recursive sequence that will get added to as
the underlying state changes, and hence writing a modification to the sequence
will result in new items being added to the sequence that are also modified. The
implications for that haven't been thought through fully yet. But I feel like I
am nearing the point where having some implementation would be helpful, to start
to iron out the inconsistencies in the spec and see whether I have actually
written any kind of consistent language!